[kernel/cpu] splitted the reaction handlers into separate definition and implementation files
