+++ kernels/cpu/src/programs/reactions/FilteredGillespieParallel.cpp
@@ -1,0 +1,9 @@
+/**
+ * << detailed description >>
+ *
+ * @file FilteredGillespieParallel.cpp
+ * @brief << brief description >>
+ * @author clonker
+ * @date 20.10.16
+ */
+
