+++ kernels/cpu/test/CPUTestReactions.cpp
@@ -10,7 +10,8 @@
 #include <gtest/gtest.h>
 #include <readdy/model/Kernel.h>
 #include <readdy/plugin/KernelProvider.h>
-#include <readdy/kernel/cpu/programs/Reactions.h>
+#include <readdy/kernel/cpu/CPUKernel.h>
+#include <readdy/kernel/cpu/programs/reactions/GillespieParallel.h>
 
 struct fix_n_threads {
     fix_n_threads(readdy::kernel::cpu::CPUKernel *const kernel, unsigned int n)
